first commit related to splitting code in smaller chunks
