+++ kcwidrp/configs/framework.cfg
@@ -44,7 +44,7 @@ logger_config_file = "logger.cfg"
 #
 pipeline_path = "", "pipelines", "examples.pipelines"
 
-primitive_path = "", "primitives"
+primitive_path = "kcwidrp.primitives", ""
 
 #
 # File type for Data_set
