+++ kcwidrp/primitives/ApplyFlat.py
@@ -1,0 +1,20 @@
+from keckdrpframework.primitives.base_primitive import BasePrimitive
+from keckdrpframework.models.arguments import Arguments
+
+
+class ApplyFlat(BasePrimitive):
+
+    def __init__(self, action, context):
+        BasePrimitive.__init__(self, action, context)
+        self.logger = context.pipeline_logger
+
+    def _perform(self):
+        self.logger.info("Applying flat field (not yet implemented)")
+
+        logstr = ApplyFlat.__module__ + "." + ApplyFlat.__qualname__
+        self.action.args.ccddata.header['HISTORY'] = logstr
+        self.logger.info(logstr)
+
+        return self.action.args
+    # END: class ApplyFlat()
+
