+++ kcwidrp/primitives/FluxCalibrate.py
@@ -1,0 +1,20 @@
+from keckdrpframework.primitives.base_primitive import BasePrimitive
+from keckdrpframework.models.arguments import Arguments
+
+
+class FluxCalibrate(BasePrimitive):
+
+    def __init__(self, action, context):
+        BasePrimitive.__init__(self, action, context)
+        self.logger = context.pipeline_logger
+
+    def _perform(self):
+        self.logger.info("Calibrating object flux (not yet implemented)")
+
+        logstr = FluxCalibrate.__module__ + "." + FluxCalibrate.__qualname__
+        self.action.args.ccddata.header['HISTORY'] = logstr
+        self.logger.info(logstr)
+
+        return self.action.args
+    # END: class FluxCalibrate()
+
