+++ kcwidrp/primitives/MakeMasterDark.py
@@ -1,0 +1,5 @@
+from keckdrpframework.primitives.base_primitive import BasePrimitive
+from keckdrpframework.models.arguments import Arguments
+
+
+
