+++ kcwidrp/primitives/ProcessObject.py
@@ -1,0 +1,12 @@
+from keckdrpframework.primitives.base_primitive import BasePrimitive
+from keckdrpframework.models.arguments import Arguments
+
+
+class ProcessObject(BasePrimitive):
+
+    def __init__(self, action, context):
+        BasePrimitive.__init__(self, action, context)
+        self.logger = context.pipeline_logger
+
+    def _perform(self):
+        return self.action.args
