+++ kcwidrp/primitives/SubtractScatteredLight.py
@@ -1,0 +1,86 @@
+from keckdrpframework.primitives.base_primitive import BasePrimitive
+from keckdrpframework.models.arguments import Arguments
+
+
+class SubtractScatteredLight(BasePrimitive):
+    """Subtract scattered light between slices"""
+
+    def __init__(self, action, context):
+        BasePrimitive.__init__(self, action, context)
+        self.logger = context.pipeline_logger
+
+    def _perform(self):
+
+        # Header keyword to update
+        key = 'SCATSUB'
+        keycom = "was scattered light subtracted?"
+
+        # Skip if nod-and-shuffle
+        if self.action.args.nasmask:
+            self.logger.info("NAS Mask: skipping scattered light subtraction")
+            self.action.args.ccddata.header[key] = (False, keycom)
+        elif self.config.instrument.skipscat:
+            self.logger.info("Skipping scattered light subtraction by request")
+            self.action.args.ccddata.header[key] = (False, keycom)
+        else:
+            # Get size of image
+            siz = self.action.args.ccddata.data.shape
+            # Get x range for scattered light
+            x0 = int(siz[1] / 2 - 180 / self.action.args.xbinsize)
+            x1 = int(siz[1] / 2 + 180 / self.action.args.xbinsize)
+            # Get y limits
+            y0 = 0
+            # y1 = int(siz[0] / 2 - 1)
+            # y2 = y1 + 1
+            y3 = siz[0]
+            # print("x limits: %d, %d, y limits: %d, %d" % (x0, x1, y0, y3))
+            # Y data values
+            yvals = np.nanmedian(self.action.args.ccddata.data[y0:y3, x0:x1],
+                                 axis=1)
+            # X data values
+            xvals = np.arange(len(yvals), dtype=np.float)
+            # Break points
+            nbkpt = int(siz[1] / 40.)
+            bkpt = xvals[nbkpt:-nbkpt:nbkpt]
+            # B-spline fit
+            bspl = sp.interpolate.LSQUnivariateSpline(xvals, yvals, bkpt)
+            if self.config.instrument.plot_level >= 1:
+                # plot
+                p = figure(title=self.action.args.plotlabel +
+                           "SCATTERED LIGHT FIT",
+                           x_axis_label='y pixel', y_axis_label='e-',
+                           plot_width=self.config.instrument.plot_width,
+                           plot_height=self.config.instrument.plot_height)
+                p.circle(xvals, yvals, legend="Scat")
+                xx = np.linspace(0, max(xvals), len(yvals) * 5)
+                p.line(xx, bspl(xx), color='red', line_width=3, legend="fit")
+                bokeh_plot(p)
+                if self.config.instrument.plot_level >= 2:
+                    input("Next? <cr>: ")
+                else:
+                    time.sleep(self.config.instrument.plot_pause)
+            # Scattered light vector
+            scat = bspl(xvals)
+            # Subtract scattered light
+            self.logger.info("Starting scattered light subtraction")
+            for ix in range(0, siz[1]):
+                self.action.args.ccddata.data[y0:y3, ix] = \
+                    self.action.args.ccddata.data[y0:y3, ix] - scat
+            self.action.args.ccddata.header[key] = (True, keycom)
+
+        logstr = SubtractScatteredLight.__module__ + \
+            "." + SubtractScatteredLight.__qualname__
+        self.action.args.ccddata.header['HISTORY'] = logstr
+        self.logger.info(logstr)
+
+        # write out int image
+        kcwi_fits_writer(self.action.args.ccddata,
+                         table=self.action.args.table,
+                         output_file=self.action.args.name, suffix="intd")
+        self.context.proctab.update_proctab(frame=self.action.args.ccddata,
+                                            suffix="intd")
+        self.context.proctab.write_proctab()
+
+        return self.action.args
+    # END: SubtractScatteredLight()
+
